+++ grakn-bootup/src/main/java/ai/grakn/bootup/BootupException.java
@@ -23,4 +23,15 @@ package ai.grakn.bootup;
  * @author Michele Orsi
  */
 public class BootupException extends RuntimeException {
+    public BootupException() {
+        super();
+    }
+
+    public BootupException(Throwable cause) {
+        super(cause);
+    }
+
+    public BootupException(String message, Throwable cause) {
+        super(message, cause);
+    }
 }
