--- grakn-bootup/src/main/java/ai/grakn/bootup/graknengine/pid/GraknPidStore.java
@@ -1,30 +1,0 @@
-/*
- * Grakn - A Distributed Semantic Database
- * Copyright (C) 2016-2018 Grakn Labs Limited
- *
- * Grakn is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Grakn is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with Grakn. If not, see <http://www.gnu.org/licenses/agpl.txt>.
- */
-
-package ai.grakn.bootup.graknengine.pid;
-
-/**
- *
- * A interface to be implemented by class who is responsible for tracking and storing the process id of Grakn
- *
- * @author Ganeshwara Herawan Hananda
- *
- */
-public interface GraknPidStore {
-    void trackGraknPid(long graknPid);
-}
